pyproject.toml
[project]
name = "otrisovka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.20.0.post0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "frozenlist==1.6.0",
    "idna==3.10",
    "magic-filter==1.0.12",
    "multidict==6.4.3",
    "pillow==11.2.1",
    "propcache==0.3.1",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "yarl==1.20.0",
]
